@@ -14,15 +14,15 @@
         "secure": "node src/server.js &! REACT_APP_SECURE=true react-scripts start"
     },
     "dependencies": {
-        "@ironcorelabs/ironweb": "^2.0.1",
-        "emotion": "^10.0.9",
-        "es6-promise": "^4.2.6",
+        "@ironcorelabs/ironweb": "^3.2.1",
+        "emotion": "^10.0.23",
+        "es6-promise": "^4.2.8",
         "jsonwebtoken": "^8.5.1",
-        "react": "16.8.6",
-        "react-dom": "16.8.6",
-        "react-redux": "^7.0.3",
-        "react-scripts": "3.0.0",
-        "redux": "^4.0.1",
+        "react": "16.11.0",
+        "react-dom": "16.11.0",
+        "react-redux": "^7.1.3",
+        "react-scripts": "3.2.0",
+        "redux": "^4.0.4",
         "redux-thunk": "^2.3.0",
         "whatwg-fetch": "^3.0.0"
     },
